All StringData usage works, UTF-16 characters fully supported. All unit tests passes both 32 and 64 bit.  Errors in Group unit tests can be replicated in pure c++ code so is not a C# issue. Added TestStuff c++ project in native solution to have a template to create c++ projects to check if a problem is on the c++ side or the c# side. Fixed some type errors in table/view sumfloat, it now returns double, not float. type errors made a unit test not work in 32bit, but work in 64bit.  Write handle pointers to console as hex when debugging.

+++ TightDbCSharp/TableOrView.cs
@@ -59,7 +59,7 @@ namespace TightDbCSharp
         internal abstract long CountDoubleNoCheck(long columnIndex, Double target);
 
         internal abstract long SumLongNoCheck(long columnIndex);
-        internal abstract float SumFloatNoCheck(long columnIndex);
+        internal abstract double SumFloatNoCheck(long columnIndex);
         internal abstract double SumDoubleNoCheck(long columnIndex);
 
         internal abstract long MinimumLongNoCheck(long columnIndex);
@@ -1232,7 +1232,7 @@ namespace TightDbCSharp
             ValidateColumnIndex(columnIndex);
             return SumLongNoCheck(columnIndex);
         }
-        public float SumFloat(long columnIndex)
+        public double SumFloat(long columnIndex)
         {
             ValidateColumnIndex(columnIndex);
             return SumFloatNoCheck(columnIndex);            
@@ -1322,7 +1322,7 @@ namespace TightDbCSharp
             long columnIndex = GetColumnIndex(columnName);
             return SumLongNoCheck(columnIndex);
         }
-        public float SumFloat(string columnName)
+        public double SumFloat(string columnName)
         {
             long columnIndex = GetColumnIndex(columnName);
             return SumFloatNoCheck(columnIndex);
